requirements.txt
streamlit
astrapy
langflow
python-dotenv
pandas
groq